Add custom generator support to open api spec (#912)
* Add custom generator support to open api spec

* Add documentation and improve naming

* fix phpdoc

* fix override generator

* Resolve conflicts

---------
