+++ src/Writing/Writer.php
@@ -141,14 +141,7 @@ class Writer
     {
         /** @var OpenAPISpecWriter $writer */
         $writer = app()->makeWith(OpenAPISpecWriter::class, ['config' => $this->config]);
-
         $spec = $writer->generateSpecContent($groupedEndpoints);
-        $overrides = $this->config->get('openapi.overrides', []);
-        if (count($overrides)) {
-            foreach ($overrides as $key => $value) {
-                data_set($spec, $key, $value);
-            }
-        }
         return Yaml::dump($spec, 20, 2, Yaml::DUMP_EMPTY_ARRAY_AS_SEQUENCE | Yaml::DUMP_OBJECT_AS_MAP);
     }
